backend/pyproject.toml
[tool.poetry]
name = "sephora-backend"
version = "0.1.0"
description = "Sephora Backend API for conversational AI assistant and reporting tool"
authors = ["Sephora Team"]
readme = "README.md"
packages = [{include = "src"}]

[project]
name = "sephora-backend"
version = "0.1.0"
description = "Sephora Backend API for conversational AI assistant and reporting tool"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "fastapi>=0.116.1",
    "google-adk>=1.12.0",
    "lxml>=6.0.1",
    "pandas>=2.3.2",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.35.0",
    "loguru (>=0.7.3,<0.8.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = ">=4.13.5"
fastapi = ">=0.116.1"
google-adk = ">=1.12.0"
lxml = ">=6.0.1"
pandas = ">=2.3.2"
requests = ">=2.32.5"
uvicorn = {extras = ["standard"], version = ">=0.35.0"}
loguru = "^0.7.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [tool.poetry.scripts]
# Scripts removed due to entry point issues - use direct commands instead:
# poetry run uvicorn src.app:app --host 0.0.0.0 --port 8000 --reload

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]